Move the auth relay classes into their own package for modularity
AuthRelayController, AuthRelayTokenService, AuthRelayException, and
Origins now live under drinkcounter.authentication.relay, separating the
cross-environment relay mechanism from the general Google auth code
(GoogleIdentityLinkingService, CustomOAuth2UserService) that lives
alongside it in drinkcounter.authentication.
