+++ src/main/java/drinkcounter/authentication/GoogleOneTapController.java
@@ -4,6 +4,8 @@ import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
 import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.api.client.json.gson.GsonFactory;
+import drinkcounter.authentication.relay.AuthRelayTokenService;
+import drinkcounter.authentication.relay.Origins;
 import drinkcounter.model.User;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
